move legacy get_zarr_store and get_mode to gefs specific utils file
