+++ tests/noaa/gefs/test_gefs_utils.py
@@ -5,7 +5,7 @@ import zarr
 from pytest import MonkeyPatch
 
 from reformatters.common.config import Config, Env
-from reformatters.common.zarr import get_mode, get_zarr_store
+from reformatters.noaa.gefs.gefs_utils import get_mode, get_zarr_store
 
 
 @pytest.mark.skip_set_local_zarr_store_base_path
